slice(blorktools): header and landing page beginnings
