+++ packages/blorktools/src/asset_debugger/index.html
@@ -6,7 +6,7 @@
     <title>Asset Debugger - Blorktools</title>
     <script>
         // Redirect to the scene directory
-        window.location.href = './scene/asset_debugger.html';
+        window.location.href = './landing-page/landing-page.html';
     </script>
     <style>
         body {
@@ -31,7 +31,7 @@
     <div class="loading">
         <h1>Asset Debugger</h1>
         <p>Redirecting to the Asset Debugger tool...</p>
-        <p>If you are not redirected automatically, <a href="./scene/asset_debugger.html" style="color: #5d8fff;">click here</a>.</p>
+        <p>If you are not redirected automatically, <a href="./landing-page/landing-page.html" style="color: #5d8fff;">click here</a>.</p>
     </div>
 </body>
 </html>
