+++ packages/blorktools/src/asset_debugger/landing-page/landing-page.html
@@ -1,0 +1,116 @@
+<!DOCTYPE html>
+<html lang="en" class="dark-mode">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Upload Section - Asset Debugger</title>
+    <link rel="stylesheet" href="../styles.css">
+    <script>
+        // Get pinned state from localStorage before DOM is constructed
+        function getPinnedState() {
+            try {
+                const savedSettings = localStorage.getItem('assetDebuggerSettings');
+                if (savedSettings) {
+                    const settings = JSON.parse(savedSettings);
+                    if (settings && settings.hasOwnProperty('pinned')) {
+                        return settings.pinned;
+                    }
+                }
+            } catch (e) {
+                console.error('Error checking pin state:', e);
+            }
+            return true; // Default to pinned if no setting or error
+        }
+        
+        // Store for later use in JS
+        window.initialPinState = getPinnedState();
+    </script>
+</head>
+<body>
+    <!-- Header container - will be populated by the header-loader.js script -->
+    <div id="header-container"></div>
+    
+    <!-- Upload section overlay -->
+    <div id="upload-section" class="tool-panel dropzone-container-main">
+        <h2 class="section-heading">Asset Files</h2>
+        <div class="dropzone-row">
+            <div class="dropzone-container">
+                <div class="dropzone" id="basecolor-dropzone">
+                    <h3>Base Color Atlas</h3>
+                    <p>Drag & drop your base color texture atlas here</p>
+                    <p class="file-info" id="basecolor-info"></p>
+                    <div class="preview" id="basecolor-preview"></div>
+                    <div class="preview-placeholder"></div>
+                </div>
+                
+                <div class="dropzone" id="orm-dropzone">
+                    <h3>ORM Atlas</h3>
+                    <p>Drag & drop your ORM (Occlusion, Roughness, Metalness) texture atlas here</p>
+                    <p class="file-info" id="orm-info"></p>
+                    <div class="preview" id="orm-preview"></div>
+                    <div class="preview-placeholder"></div>
+                </div>
+                
+                <div class="dropzone" id="normal-dropzone">
+                    <h3>Normal Atlas</h3>
+                    <p>Drag & drop your normal map texture atlas here</p>
+                    <p class="file-info" id="normal-info"></p>
+                    <div class="preview" id="normal-preview"></div>
+                    <div class="preview-placeholder"></div>
+                </div>
+                
+                <div class="dropzone" id="model-dropzone">
+                    <h3>3D Model</h3>
+                    <p>Drag & drop a GLB model file here</p>
+                    <p>If not provided, a cube will be used</p>
+                    <p class="file-info" id="model-info"></p>
+                    <div class="preview" id="model-preview"></div>
+                    <div class="preview-placeholder"></div>
+                </div>
+            </div>
+        </div>
+        
+        <h2 class="section-heading">World Files</h2>
+        <div class="dropzone-row">
+            <div class="dropzone-container" style="display: grid; grid-template-columns: repeat(4, 1fr); gap: 15px;">
+                <div style="grid-column: 1/2; visibility: hidden;"></div>
+                
+                <div class="dropzone" id="lighting-dropzone" style="grid-column: 2/3;">
+                    <h3>Lighting File</h3>
+                    <p>Drag & drop your HDR or EXR lighting file here</p>
+                    <p class="file-info" id="lighting-info"></p>
+                    <div class="preview" id="lighting-preview"></div>
+                    <div class="preview-placeholder"></div>
+                </div>
+                
+                <div class="dropzone" id="background-dropzone" style="grid-column: 3/4;">
+                    <h3>Background Image</h3>
+                    <p>Drag & drop your HDR, EXR, JPEG, PNG, WebP, or TIFF background image here</p>
+                    <p class="file-info" id="background-info"></p>
+                    <div class="preview" id="background-preview"></div>
+                    <div class="preview-placeholder"></div>
+                </div>
+                
+                <div style="grid-column: 4/5; visibility: hidden;"></div>
+            </div>
+        </div>
+        
+        <p id="zip-info" class="file-info" style="text-align: center; margin-top: 15px; display: none;"></p>
+        
+        <button id="start-debug" class="start-button">Start Debugging</button>
+    </div>
+    
+    <!-- Load the header component -->
+    <script type="module">
+        import { loadHeader } from '../header/header-loader.js';
+        
+        // Load the header when the DOM is ready
+        document.addEventListener('DOMContentLoaded', () => {
+            loadHeader('header-container');
+        });
+    </script>
+    
+    <!-- Main application script -->
+    <script type="module" src="./landing-page.js"></script>
+</body>
+</html>
