+++ packages/blorktools/src/asset_debugger/landing-page/landing-page.js
@@ -1,0 +1,7 @@
+import { initState } from "../scene/state";
+import { setupDropzones } from "./dropzone-util";
+
+export function initalizeLandingPage() {
+    initState();
+    setupDropzones();
+}
